moved annotations, added @WithExample
