+++ src/main/kotlin/com/papsign/ktor/openapigen/annotations/type/number/floating/FloatingNumberConstraintProcessor.kt
@@ -1,0 +1,9 @@
+package com.papsign.ktor.openapigen.annotations.type.number.floating
+
+import com.papsign.ktor.openapigen.getKType
+import com.papsign.ktor.openapigen.annotations.type.number.NumberConstraintProcessor
+
+abstract class FloatingNumberConstraintProcessor<A: Annotation>: NumberConstraintProcessor<A>(listOf(
+    getKType<Float>(),
+    getKType<Double>()
+))
