+++ src/main/kotlin/com/papsign/ktor/openapigen/annotations/type/number/floating/clamp/FClamp.kt
@@ -1,0 +1,11 @@
+package com.papsign.ktor.openapigen.annotations.type.number.floating.clamp
+
+import com.papsign.ktor.openapigen.schema.processor.SchemaProcessorAnnotation
+import com.papsign.ktor.openapigen.validation.ValidatorAnnotation
+
+@Target(AnnotationTarget.TYPE, AnnotationTarget.PROPERTY)
+@SchemaProcessorAnnotation(FClampProcessor::class)
+@ValidatorAnnotation(FClampProcessor::class)
+annotation class FClamp(val min: Double, val max: Double)
+
+
