+++ src/main/kotlin/com/papsign/ktor/openapigen/annotations/type/number/floating/min/FMin.kt
@@ -1,0 +1,9 @@
+package com.papsign.ktor.openapigen.annotations.type.number.floating.min
+
+import com.papsign.ktor.openapigen.schema.processor.SchemaProcessorAnnotation
+import com.papsign.ktor.openapigen.validation.ValidatorAnnotation
+
+@Target(AnnotationTarget.TYPE, AnnotationTarget.PROPERTY)
+@SchemaProcessorAnnotation(FMinProcessor::class)
+@ValidatorAnnotation(FMinProcessor::class)
+annotation class FMin(val value: Double)
