+++ src/main/kotlin/com/papsign/ktor/openapigen/annotations/type/number/integer/IntegerNumberConstraintProcessor.kt
@@ -1,0 +1,11 @@
+package com.papsign.ktor.openapigen.annotations.type.number.integer
+
+import com.papsign.ktor.openapigen.getKType
+import com.papsign.ktor.openapigen.annotations.type.number.NumberConstraintProcessor
+
+abstract class IntegerNumberConstraintProcessor<A: Annotation>: NumberConstraintProcessor<A>(listOf(
+    getKType<Int>(),
+    getKType<Long>(),
+    getKType<Float>(),
+    getKType<Double>()
+))
