+++ src/main/kotlin/com/papsign/ktor/openapigen/annotations/type/number/integer/clamp/Clamp.kt
@@ -1,0 +1,9 @@
+package com.papsign.ktor.openapigen.annotations.type.number.integer.clamp
+
+import com.papsign.ktor.openapigen.schema.processor.SchemaProcessorAnnotation
+import com.papsign.ktor.openapigen.validation.ValidatorAnnotation
+
+@Target(AnnotationTarget.TYPE, AnnotationTarget.PROPERTY)
+@SchemaProcessorAnnotation(ClampProcessor::class)
+@ValidatorAnnotation(ClampProcessor::class)
+annotation class Clamp(val min: Long, val max: Long)
