+++ src/main/kotlin/com/papsign/ktor/openapigen/annotations/type/number/integer/max/Max.kt
@@ -1,0 +1,9 @@
+package com.papsign.ktor.openapigen.annotations.type.number.integer.max
+
+import com.papsign.ktor.openapigen.schema.processor.SchemaProcessorAnnotation
+import com.papsign.ktor.openapigen.validation.ValidatorAnnotation
+
+@Target(AnnotationTarget.TYPE, AnnotationTarget.PROPERTY)
+@SchemaProcessorAnnotation(MaxProcessor::class)
+@ValidatorAnnotation(MaxProcessor::class)
+annotation class Max(val value: Long)
