+++ src/main/kotlin/com/papsign/ktor/openapigen/annotations/type/number/integer/min/Min.kt
@@ -1,0 +1,9 @@
+package com.papsign.ktor.openapigen.annotations.type.number.integer.min
+
+import com.papsign.ktor.openapigen.schema.processor.SchemaProcessorAnnotation
+import com.papsign.ktor.openapigen.validation.ValidatorAnnotation
+
+@Target(AnnotationTarget.TYPE, AnnotationTarget.PROPERTY)
+@SchemaProcessorAnnotation(MinProcessor::class)
+@ValidatorAnnotation(MinProcessor::class)
+annotation class Min(val value: Long)
