+++ src/main/kotlin/com/papsign/ktor/openapigen/annotations/type/object/example/NoExampleProvider.kt
@@ -1,0 +1,3 @@
+package com.papsign.ktor.openapigen.annotations.type.`object`.example
+
+internal class NoExampleProvider private constructor(): ExampleProvider<Any?>
