+++ src/main/kotlin/com/papsign/ktor/openapigen/annotations/type/object/example/WithExample.kt
@@ -1,0 +1,12 @@
+package com.papsign.ktor.openapigen.annotations.type.`object`.example
+
+import com.papsign.ktor.openapigen.schema.processor.SchemaProcessorAnnotation
+import kotlin.reflect.KClass
+
+/**
+ * Careful, no type checking done if you give the wrong provider
+ */
+@Target(AnnotationTarget.CLASS, AnnotationTarget.TYPE, AnnotationTarget.PROPERTY)
+@SchemaProcessorAnnotation(ExampleProcessor::class)
+annotation class WithExample(val provider: KClass<out ExampleProvider<*>> = NoExampleProvider::class)
+
