+++ src/main/kotlin/com/papsign/ktor/openapigen/annotations/type/string/lowercase/LowerCase.kt
@@ -1,0 +1,8 @@
+package com.papsign.ktor.openapigen.annotations.type.string.lowercase
+
+import com.papsign.ktor.openapigen.validation.ValidatorAnnotation
+
+@Target(AnnotationTarget.PROPERTY, AnnotationTarget.TYPE)
+@ValidatorAnnotation(LowerCaseValidator::class)
+annotation class LowerCase
+
