+++ src/main/kotlin/com/papsign/ktor/openapigen/annotations/type/string/trim/Trim.kt
@@ -1,0 +1,8 @@
+package com.papsign.ktor.openapigen.annotations.type.string.trim
+
+import com.papsign.ktor.openapigen.validation.ValidatorAnnotation
+
+@Target(AnnotationTarget.PROPERTY, AnnotationTarget.TYPE)
+@ValidatorAnnotation(TrimValidator::class)
+annotation class Trim
+
