+++ src/main/kotlin/com/papsign/ktor/openapigen/model/schema/SchemaModel.kt
@@ -5,32 +5,39 @@ import com.papsign.ktor.openapigen.model.base.RefModel
 
 sealed class SchemaModel<T>: DataModel {
 
+    abstract var example: T?
+    abstract var examples: List<T>?
+
     data class SchemaModelObj<T>(
         var properties: Map<String, SchemaModel<*>>,
         var required: List<String>,
         var nullable: Boolean = false,
-        var example: T? = null,
+        override var example: T? = null,
+        override var examples: List<T>? = null,
         var type: DataType = DataType.`object`
     ) : SchemaModel<T>()
 
     data class SchemaModelMap<T : Map<String, U>, U>(
         var additionalProperties: SchemaModel<U>,
         var nullable: Boolean = false,
-        var example: T? = null,
+        override var example: T? = null,
+        override var examples: List<T>? = null,
         var type: DataType = DataType.`object`
     ) : SchemaModel<T>()
 
     data class SchemaModelEnum<T>(
         var enum: List<String>,
         var nullable: Boolean = false,
-        var example: T? = null,
+        override var example: T? = null,
+        override var examples: List<T>? = null,
         var type: DataType = DataType.string
     ) : SchemaModel<T>()
 
     data class SchemaModelArr<T>(
         var items: SchemaModel<*>?,
         var nullable: Boolean = false,
-        var example: T? = null,
+        override var example: T? = null,
+        override var examples: List<T>? = null,
         var type: DataType = DataType.array
     ) : SchemaModel<T>()
 
@@ -40,10 +47,17 @@ sealed class SchemaModel<T>: DataModel {
         var nullable: Boolean = false,
         var minimum: T? = null,
         var maximum: T? = null,
-        var example: T? = null
+        override var example: T? = null,
+        override var examples: List<T>? = null
     ) : SchemaModel<T>()
 
-    data class SchemaModelRef<T>(override val `$ref`: String) : SchemaModel<T>(), RefModel<SchemaModel<T>>
+    data class SchemaModelRef<T>(override val `$ref`: String) : SchemaModel<T>(), RefModel<SchemaModel<T>> {
+        override var example: T? = null
+        override var examples: List<T>? = null
+    }
 
-    data class OneSchemaModelOf<T>(val oneOf: List<SchemaModel<out T>>) : SchemaModel<T>()
+    data class OneSchemaModelOf<T>(val oneOf: List<SchemaModel<out T>>) : SchemaModel<T>() {
+        override var example: T? = null
+        override var examples: List<T>? = null
+    }
 }
