--- src/main/kotlin/com/papsign/ktor/openapigen/validation/number/floating/FloatingNumberConstraintProcessor.kt
@@ -1,9 +1,0 @@
-package com.papsign.ktor.openapigen.validation.number.floating
-
-import com.papsign.ktor.openapigen.getKType
-import com.papsign.ktor.openapigen.validation.number.NumberConstraintProcessor
-
-abstract class FloatingNumberConstraintProcessor<A: Annotation>: NumberConstraintProcessor<A>(listOf(
-    getKType<Float>(),
-    getKType<Double>()
-))
