--- src/main/kotlin/com/papsign/ktor/openapigen/validation/number/floating/max/FMax.kt
@@ -1,9 +1,0 @@
-package com.papsign.ktor.openapigen.validation.number.floating.max
-
-import com.papsign.ktor.openapigen.schema.processor.SchemaProcessorAnnotation
-import com.papsign.ktor.openapigen.validation.ValidatorAnnotation
-
-@Target(AnnotationTarget.TYPE)
-@SchemaProcessorAnnotation(FMaxProcessor::class)
-@ValidatorAnnotation(FMaxProcessor::class)
-annotation class FMax(val value: Double)
