--- src/main/kotlin/com/papsign/ktor/openapigen/validation/number/floating/min/FMin.kt
@@ -1,9 +1,0 @@
-package com.papsign.ktor.openapigen.validation.number.floating.min
-
-import com.papsign.ktor.openapigen.schema.processor.SchemaProcessorAnnotation
-import com.papsign.ktor.openapigen.validation.ValidatorAnnotation
-
-@Target(AnnotationTarget.TYPE)
-@SchemaProcessorAnnotation(FMinProcessor::class)
-@ValidatorAnnotation(FMinProcessor::class)
-annotation class FMin(val value: Double)
