--- src/main/kotlin/com/papsign/ktor/openapigen/validation/string/trim/Trim.kt
@@ -1,8 +1,0 @@
-package com.papsign.ktor.openapigen.validation.string.trim
-
-import com.papsign.ktor.openapigen.validation.ValidatorAnnotation
-
-@Target(AnnotationTarget.PROPERTY, AnnotationTarget.TYPE)
-@ValidatorAnnotation(TrimValidator::class)
-annotation class Trim
-
